+++ hypervisor/Makefile
@@ -219,6 +219,7 @@ HW_C_SRCS += arch/x86/gdt.c
 HW_C_SRCS += arch/x86/nmi.c
 HW_C_SRCS += arch/x86/exception.c
 HW_C_SRCS += arch/x86/irq.c
+HW_C_SRCS += arch/x86/tsc.c
 HW_C_SRCS += arch/x86/timer.c
 HW_C_SRCS += arch/x86/vmx.c
 HW_C_SRCS += arch/x86/cpu_state_tbl.c
@@ -228,6 +229,7 @@ HW_C_SRCS += arch/x86/trampoline.c
 HW_S_SRCS += arch/x86/sched.S
 HW_C_SRCS += arch/x86/rdt.c
 HW_C_SRCS += arch/x86/sgx.c
+HW_C_SRCS += common/ticks.c
 HW_C_SRCS += common/irq.c
 HW_C_SRCS += common/softirq.c
 HW_C_SRCS += common/schedule.c
