+++ hypervisor/debug/profiling.c
@@ -15,6 +15,7 @@
 #include <asm/guest/vm.h>
 #include <sprintf.h>
 #include <logmsg.h>
+#include <ticks.h>
 
 #define DBG_LEVEL_PROFILING		5U
 #define DBG_LEVEL_ERR_PROFILING		3U
@@ -339,7 +340,7 @@ static int32_t profiling_generate_data(int32_t collector, uint32_t type)
 			clac();
 
 			/* populate the data header */
-			pkt_header.tsc = rdtsc();
+			pkt_header.tsc = cpu_ticks();
 			pkt_header.collector_id = collector;
 			pkt_header.cpu_id = get_pcpu_id();
 			pkt_header.data_type = 1U << type;
@@ -411,7 +412,7 @@ static int32_t profiling_generate_data(int32_t collector, uint32_t type)
 		clac();
 
 		/* populate the data header */
-		pkt_header.tsc = rdtsc();
+		pkt_header.tsc = cpu_ticks();
 		pkt_header.collector_id = collector;
 		pkt_header.cpu_id = get_pcpu_id();
 		pkt_header.data_type = (uint16_t)type;
@@ -1307,7 +1308,7 @@ void profiling_vmenter_handler(__unused struct acrn_vcpu *vcpu)
 			((socwatch_collection_switch &
 				(1UL << (uint64_t)SOCWATCH_VM_SWITCH_TRACING)) > 0UL))) {
 
-		get_cpu_var(profiling_info.vm_info).vmenter_tsc = rdtsc();
+		get_cpu_var(profiling_info.vm_info).vmenter_tsc = cpu_ticks();
 	}
 }
 
@@ -1323,7 +1324,7 @@ void profiling_pre_vmexit_handler(struct acrn_vcpu *vcpu)
 	if ((get_cpu_var(profiling_info.s_state).pmu_state == PMU_RUNNING) ||
 		(get_cpu_var(profiling_info.soc_state) == SW_RUNNING)) {
 
-		get_cpu_var(profiling_info.vm_info).vmexit_tsc = rdtsc();
+		get_cpu_var(profiling_info.vm_info).vmexit_tsc = cpu_ticks();
 		get_cpu_var(profiling_info.vm_info).vmexit_reason
 			= exit_reason;
 		if (exit_reason == VMX_EXIT_REASON_EXTERNAL_INTERRUPT) {
