+++ hypervisor/debug/trace.c
@@ -6,6 +6,7 @@
 
 #include <types.h>
 #include <asm/per_cpu.h>
+#include <ticks.h>
 #include <trace.h>
 
 #define TRACE_CUSTOM			0xFCU
@@ -51,7 +52,7 @@ static inline void trace_put(uint16_t cpu_id, uint32_t evid, uint32_t n_data, st
 {
 	struct shared_buf *sbuf = per_cpu(sbuf, cpu_id)[ACRN_TRACE];
 
-	entry->tsc = rdtsc();
+	entry->tsc = cpu_ticks();
 	entry->id = evid;
 	entry->n_data = (uint8_t)n_data;
 	entry->cpu = (uint8_t)cpu_id;
