+++ hypervisor/include/arch/x86/asm/timer.h
@@ -8,6 +8,7 @@
 #define TIMER_H
 
 #include <list.h>
+#include <asm/tsc.h>
 
 /**
  * @brief Timer
@@ -47,37 +48,7 @@ struct hv_timer {
 
 /* External Interfaces */
 
-#define CYCLES_PER_MS	us_to_ticks(1000U)
-
 void udelay(uint32_t us);
-
-/**
- * @brief convert us to ticks.
- *
- * @return ticks
- */
-uint64_t us_to_ticks(uint32_t us);
-
-/**
- * @brief convert ticks to us.
- *
- * @return microsecond
- */
-uint64_t ticks_to_us(uint64_t ticks);
-
-/**
- * @brief convert ticks to ms.
- *
- * @return millisecond
- */
-uint64_t ticks_to_ms(uint64_t ticks);
-
-/**
- * @brief read tsc.
- *
- * @return tsc value
- */
-uint64_t rdtsc(void);
 
 /**
  * @brief Initialize a timer structure.
@@ -161,20 +132,6 @@ void del_timer(struct hv_timer *timer);
  * @return None
  */
 void timer_init(void);
-
-/**
- * @brief Calibrate tsc.
- *
- * @return None
- */
-void calibrate_tsc(void);
-
-/**
- * @brief  Get tsc.
- *
- * @return tsc(KHz)
- */
-uint32_t get_tsc_khz(void);
 
 /**
  * @}
